Test fixes

+++ audiotrails/episodes/views/history.py
@@ -1,4 +1,7 @@
+import http
+
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import redirect
 from django.views.decorators.http import require_POST
 
@@ -35,6 +38,10 @@ def index(request):
 @ajax_login_required
 def remove_audio_log(request, episode_id):
     episode = get_episode_or_404(request, episode_id)
+
+    # you shouldn't be able to remove history if episode currently playing
+    if request.player.is_playing(episode):
+        return HttpResponse(status=http.HTTPStatus.NO_CONTENT)
 
     AudioLog.objects.filter(user=request.user, episode=episode).delete()
